split up macros
